requirements.txt
Flask==2.3.3
pymongo==4.4.0
python-dotenv==1.0.0
Werkzeug==3.0.0
